requirements.txt
discord.py
python-dotenv
ollama
requests
numpy
tiktoken